I think I figured it out this time... :) Its a decimal value... Not float... :)

+++ cs50-php-app/src/Model/Table/SensordataTable.php
@@ -49,7 +49,7 @@ class SensordataTable extends Table
             ->allowEmpty('id', 'create');
 
         $validator
-            ->float('temp1')
+            ->decimal('temp1')
             ->requirePresence('temp1', 'create')
             ->notEmpty('temp1');
